fix: split out buf tests to allow for skipping without a marker
