+++ pyproject.toml
@@ -59,7 +59,6 @@ addopts = [
   "--ignore=.direnv",
   "--strict-markers"
 ]
-markers = ["buf"]
 norecursedirs = ["site-packages", "dist-packages", ".direnv"]
 
 [tool.poetry2conda]
